15.websites/04.tesseract/requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
Pillow==10.0.0
pdf2image==1.16.3
